% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lib.R
\name{notes}
\alias{notes}
\title{notes}
\usage{
notes(path = NULL, json = NULL)
}
\arguments{
\item{path}{path to destination file.}

\item{json}{source json}
}
\description{
Create zotero notes and open them.
}
